refactor(distribution): fixed re-exporting and removed peer dependancy
