+++ package.json
@@ -33,7 +33,6 @@
     "version": "node -v",
     "browsers": "testcafe --list-browsers",
     "semantic-release": "semantic-release",
-    "install:peers": "npm-install-peers",
     "pack:contents": "npm pack && tar -xvzf *.tgz && rm -rf package *.tgz"
   },
   "config": {
@@ -44,10 +43,8 @@
   "files": [
     "dist/*"
   ],
-  "peerDependencies": {
-    "allure-js-commons": "^2.0.0-beta.6"
-  },
   "dependencies": {
+    "allure-js-commons": "^2.0.0-beta.6",
     "merge-anything": "^3.0.3",
     "rimraf": "^3.0.2"
   },
@@ -69,7 +66,6 @@
     "jest": "^25.3.0",
     "lint-staged": "^10.1.3",
     "normalize-newline": "^3.0.0",
-    "npm-install-peers": "^1.2.1",
     "prettier": "^2.0.4",
     "semantic-release": "^17.0.4",
     "testcafe": "^1.8.4",
